bypass creating an api connection when the help parameter is given
By default when a commands sub-command is invoked, it will create an
api connection to the MIQ endpoint. This means any parameter given when
the sub-command is invoked would attempt to connect to the endpoint. It
should not for the help parameter since its purpose is to show the
available parameters and exit. This patch resolves connecting when help
is given.