improve CLientAPI constructor
By improving the constructor, we protect the class to misbehave
in case settings is changed by an external entity. This commit
also improves the connect function by breaking it down in smaller
functions that do one thing only.

In addition to this commit, it also disable urllib3 warnings about
connecting to an https instance with an invalid certificate.

Closes #45