fix logging debug messages within sub-commands
This commit resolves a bug found when trying to log debug level
messages within the sub-command layer of the CLI. Originally it was
getting the current context and trying to see if verbose mode was set.
It was getting the context for the sub-command and not the root context.
The root context is the one that allows you to enable verbose
mode and saves its value. Now the debug function will get the
root context and not raise a KeyError. Allowing sub-commands to log
debug messages.

Closes #51