improve the config class (#56)
This commit improves the logging messages with the config class when
an error may occur. The previous design silenced every message so the
user would never know exactly why their config may not be loaded
properly. This is now handled by the verbose option the cli has. It
provides the ability to give more information back to the user if
enabled.

In order for this to work, I needed to move loading config settings
right before we invoke the command. Why? We do not have the verbose
setting until we invoke the command. Previously we loaded the config
settings right before the cli began and at that point we would never
have the value the user gives for verbose option.

With the change for moving where we load the config, it required a
change to how each collection class has access to the client api
pointer. Previously each collection class would inherit the ClientAPI
to have access to it. This was fine because we had all the config
settings at that time. Click needs the collection class object in order
to build the cli. With the config settings loaded before the command
is invoked, the client api pointer previously created would be invalid.
With the new design, right before invoking the command, we create a
client api pointer with the final config settings loaded, create the
connection to the server and save the pointer within click context.
Each collection class method has a decorator which provides the class
with the client api pointer. Providing the same ability the previous
design did when inheriting the ClientAPI class.

Closes #55 and resolves #50