add abort logging function
This commit adds a new logging function (log.abort()) to the CLI. It
should be used when you want to log and exit the CLI. The abort function
accepts a message which will then log it as an error message and exit
with a return code of non zero. Previously the log.error() would allow
you to exit the CLI but it will now just log the messages.

Also includes updated doc strings for each logging function. They
explain when it is best to use each one.

Closes #64