log warning message when default configuration is used
This commit provides the user with a helpful message when the CLI is
using the default configuration. A comparison is performed between the
default settings and the final cli parameters to determine if the
message is logged.

Removes verbose from DEFAULT_CONFIG constant. Those settings should be
strictly for the client api.

Closes #66