set useful collection attributes within the client api decorator (#68)
Previously the client_api decorator would set the collection api pointer
attribute to easily access the client api library when the method is
run. This still would require code to lookup the exact collection
object from the api pointer and that collection pointer action to be
run. This commit provides three new attributes:

1. collection: the exact collection reference.
    -> i.e. provider collection class `collection` attribute has the ref
    to the provider collection only.

2. all: all entities regarding the collection.
    -> i.e. provider collection class `all` attribute has all the
    entities regarding the provider collection.

3. action: the exact collection action reference.
    -> i.e. provider collection class `action` attribute has the ref
    to the provider action only.

This change will be used by #23