adds a class for easy use of the client in python modules (#77)
With the addition of a unified client class. Users will be able to use
the MIQ cli directly within python modules. This class eliminates
multiple steps the user would have had to perform. Please reference the
class doc string for more details on how to use.

It also relocates disabling the urllib3 warning message. Previously
this was done in the cli package but now is moved to the api module. It
is needed when using the client via command line or as a library import.

Closes #74