fix recursion error from being raised when getting action attribute (#91)
* fix recursion error from being raised when getting action attribute

This commit handles the recursion error raised by the api client
library when a collection action is not found.

Closes #89

* use RuntimeError over RecursionError

To maintain compatibility for Python 2, need to catch RuntimeError.
Python 2 does not have a built-in exception for RecursionError. In
Python 3, RecursionError inherits the RuntimeError. Using this will
handle catching recursion errors.