add provider class for easy id lookup thats transparent of the provider
This commit adds a new module containing classes regarding components
for a provider. Collections can easily create a provider object and
perform lookups on a string element to return the given id in manageiq.
The class will automatically determine the resource type to filter on.
Removing the need to have static strings in a collection class.

It also includes the following:
  * modifications to automation/provision requests collections to use
  the new provider class.

  * cosmetic changes to automation/provision requests collections.

  * automation/provision requests status previously accepted the id by
  a click option. Switched to a argument. This way users do not need to
  define the --id and can just give the id. This command only needs the
  id so it makes sense to be positional. Also removes use of built-in
  as a variable.

Closes #86