fix config load from yaml when the yaml file is wrong
From design, we want to load configuration from a yaml file. If the file
doesn't exist we want to silently carry on. But if the file exist and
there is an error in the file (i.e. yaml syntax), we want the client to
fail and let the user know that the failure occurred during
configuration loading. If verbose is turned ON, than a bit more
information will be shown.