fix loading config when config file is invalid
This commit will add an extra layer of protection when trying to parse
a config file that has invalid data. An AttributeError was seen when
trying to get key:value entries from a config file in the form of a
string data type. A check is put in place to quit the client if the
config file content is in string format. Config file format should be
in the form of a dictionary data type.