catch syntax error when loading config from environment
The module ast.literal_eval will throw a SyntaxError exception
in case the given parameter is not part of its acceptable formats
list.

In this case, like loading from yaml, we want the CLI to fail
because there's an intention of the user to use environment
variable but the content is not correctly formated.