add test coverage for multiple modules (#107)
* update testing environment for functional and integration tests

* add two options in tox: functional and integration. The functional
runs the tests located at tests/functional. Tests in that location
is for unit and functional tests, using mocks for external libraries
that talk to a ManageIQ instance.

* downgraded the docker-py requirement to 2.1.0 for compatibility
reasons (currently, core developers use the latest docker for RHEL 7.4)

* add few initial tests to test the functional tests environment.

* add 'cover/' folder into the ignore list as the tox environments
for testing will create coverage report folders for each test run.

* add 20 tests for api module

* add tests for the utils.log module

* add tests for utils module

This tests increase the coverage from 0% to 55% of the
utils module. Different types of configuration files
(including wrong ones) are added with this commit.

* fix tests and compatibility issue

* fix compatibility with Python 3 by removing the has_key `method`
  and replace with `in`

* remove a repeated test