openstack provisioning update (#113)
This commit brings updates to provision requests within openstack providers. Previously
the key pair and security group keys were required. These keys are now marked as optional.

It also adds a status method for the request task collection. This method allows you to pass a
request ID and returns back the status about the given request in process.

Openstack example scripts have been updated to use the request task updates to check the
status of a task. Instead of directly checking collection that returned the ID.