add warning if etc config directory does not exist
Currently, if the `/etc/miqcli` does not exist, an error occurs.
As per earlier decision, we do not want the client to fail if
the main configuration directory does not exist, but carry on
looking for more precedent configurations or use the default
pre-baked into the project.

This commit fixes this by throwing a warning message that the
directory does not exist and keep the client running as expected.